TIMS-1049: replacement or BLOB type columns in DATATAG and COMMAND database tables. Used VARCHAR types instead. Adaptation of related Unit tests and fix of some minor issues related to simple types casting in some of those tests.

+++ c2mon-server/c2mon-server-cachedbaccess/src/test/cern/c2mon/server/cache/dbaccess/CommandTagMapperTest.java
@@ -101,8 +101,8 @@ public class CommandTagMapperTest {
     modifiedCommand.setSourceRetries(2);
     modifiedCommand.setExecTimeout(10);
     modifiedCommand.setClientTimeout(3);
-    modifiedCommand.setMinimum(Float.valueOf(30));
-    modifiedCommand.setMaximum(Float.valueOf(60)); 
+    modifiedCommand.setMinimum(Integer.valueOf(30));
+    modifiedCommand.setMaximum(Integer.valueOf(60)); 
     
     //update
     commandTagMapper.updateCommandTag(modifiedCommand);
